+++ src/synth/oscillator.c
@@ -17,26 +17,20 @@ static uint32_t phase = 0;                  // An indicator of the current posit
 static uint32_t step = 0;                   // Step size for each sample
 
 void osc_init(void) {
-    // fill sine_table once using sin()
-    // reset phase
     for (int i = 0; i < TABLE_SIZE; i++) {
         double x = 2.0 * M_PI * (double)i / (double)TABLE_SIZE;
-        sine_table[i] = (int16_t)(AMPLITUDE * sin(x));
+        sine_table[i] = (int32_t)(AMPLITUDE * sin(x));
     }
     phase = 0;
     step = 0;
 }
 
 void osc_set_frequency(uint32_t freq_hz) {
-    // step = freq * 2^32 / SAMPLE_RATE
     step = ((uint64_t)freq_hz << 32) / SAMPLE_RATE;
 }
 
-int16_t osc_next_sample(void) {
-    // phase += step
-    // index = phase >> 24
-    // return sine_table[index]
+int32_t osc_next_sample(void) {
     phase += step;
-    uint8_t idx = phase >> 24; // get the top 8 bits for index (0-255)
+    uint8_t idx = phase >> 24;
     return sine_table[idx];
 }
